Fix/silence some warnings in 'make check'
'make check' currently fails with the following in src/test-suite.log
(how can one tell automake not to hide the error message in a file?!?):

FAIL: check-doc-syntax.sh
=========================

Checking documentation for incorrect syntax
./cairo-cogl-surface.c: * cairo_cogl_pipeline_ts. On failure, both pointers will be set to
./cairo-cogl-surface.c:/* Mostly taken from cairo_vg_surface.c */
Error: some function names in the docs are not followed by parentheses.
Fix this by searching for the following regexp in the above files:
	'^[^:]*:[/ ][*]\(\|[ 	].*\)\([^#']\|^\)\<\(cairo_[][<>/0-9a-z_]*\>[^][<>(]\)'
FAIL check-doc-syntax.sh (exit status: 1)

This commit silences that. The fix is likely not correct, but since
these are normal comments and not actual doc comments, it does not
matter much.

Signed-off-by: Uli Schlachter <psychon@znc.in>