Switch to line-separated test failure lists
Having comma separated lists of test names that we expect to fail leads
to really unreadable and useless diffs. Thus, this commit changes these
lists to be separated by newlines. This change was generated with:

  ( cd .gitlab-ci ; for x in * ; do tr ',' '\n' < $x > .$x ; mv .$x $x ; done )

The code in .gitlab-ci.yml is changed to use tr again to transfer these
lists back so that everything still works.

Signed-off-by: Uli Schlachter <psychon@znc.in>