CI: Split test execution into per-backend jobs
This creates one job per cairo backend that is tested on CI. That
backend is then tested in its own job.

One motivation is that this will hopefully be faster, because tests are
run in parallel. Another motivation is that this lets us get around the
one hour time limit per job.

A downside is that the inter-backend tests are not run in CI. For
example, xlib-surface-source requires a working X11 server. I am not
providing that X11 server here.

Signed-off-by: Uli Schlachter <psychon@znc.in>