Handle _cairo_strtod() failures in parse_float()
When strtod cannot do anything, it returns zero and sets the end pointer
to the beginning of the string. This commit changes the code in
parse_float() to treat this case as an error.

Without the fix from commit b7d67433b7c0, this commit turns the error
from an endless loop into a parse error, which is a lot better error
behaviour.

Signed-off-by: Uli Schlachter <psychon@znc.in>